website changes
Updated linking paths in the project to use the new directory structure for static assets and icons. Added a new PDF for the CV and adjusted the language icons for the resume. Made enhancements to the contact info in the resume section and added hover effects for social links. Reviewed and optimized the Nginx configuration for the website.