this commit is filled with a bunch of wierd hacks and workarounds
but it sort of works.

now window switching also works in conjunction with neovim, and the window manager.
and there is now a plugin that allows issuing commands to a neovim
instance based on the PID. and the PID is written to the title of the
window. allowing it to be parsed and found